requirements.txt
appnope==0.1.0
asgiref==3.2.10
autopep8==1.5.4
backcall==0.2.0
boto3==1.16.3
botocore==1.19.3
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.7
django-admin-sortable2==0.7.6
django-angular==2.3
django-appconf==1.0.4
django-ckeditor==6.0.0
django-classy-tags==1.0.0
django-cms==3.7.3
django-cors-headers==3.4.0
django-entangled==0.3
django-environ==0.4.5
django-extensions==2.2.9
django-filer==1.7.1
django-formtools==2.2
django-fsm==2.7.0
django-fsm-admin==1.2.4
django-heroku==0.3.1
django-ipware==2.1.0
django-jazzmin==2.2.6
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
django-post-office==3.4.1
django-rename-app==0.1.2
django-rest-auth==0.9.5
django-sekizai==1.1.0
django-select2==7.4.2
django-shop==1.2.2
django-storages==1.10.1
django-templated-mail==1.1.1
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-cascade==1.3.2
djangocms-text-ckeditor==3.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.0.3
easy-thumbnails==2.7
Faker==4.1.1
gunicorn==20.0.4
html5lib==1.1
httplib2==0.18.1
idna==2.10
install==1.3.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
jmespath==0.10.0
jsonfield==3.1.0
oauth2client==4.1.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.5
psycopg2==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydot==1.4.1
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.1
rsa==4.6
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
stripe==2.55.1
text-unidecode==1.3
toml==0.10.2
traitlets==4.3.3
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
